Refactor NotificationContext into separate modules
Extract NotificationContext type into its own file and reorganize related logic. Split `useNotification` into a dedicated module and update imports across components for better modularity and maintainability.
